fix: undo emotion-related changes
