+++ packages/app-website/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./createApolloClient";
+export * from "./createEmotionCache";
 export * from "./isPrerendering";
 export * from "./getPrerenderId";
